requirements.txt
asgiref==3.7.2
Django~=4.2.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3

pip~=22.0.2
wheel~=0.37.1
Brotli~=1.0.9
cryptography~=3.4.8
Pillow~=9.0.1
keyring~=23.5.0
colorama~=0.4.4
PyYAML~=5.4.1
requests~=2.25.1
click~=8.0.3
setuptools~=59.6.0
future~=0.18.2
python-dotenv~=1.0.0